(MultipleFiles) - refactor: folders
