+++ front/src/components/templates/PageTemplate.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SideBar from 'components/organisms/SideBar/SideBar';
+import SideBar from 'components/organisms/navigation/SideBar/SideBar';
 import LoginButtons from 'components/molecules/clickable/LoginButtons/LoginButtons';
 
 const PageTemplate = ({children}) => {
